Configure Input Groups in mythtv-setup
Capture cards can be configured so that only one capture card of a
group of capture cards can be used at any one time. This is done
by defining an Input Group and then assigning that group to all
capture cards that are to be part of the group.
Configuring this with mythtv-setup has always been possible but there
was a bug that allowed only one card to be configured as member of an
input group. This bug is now fixed.

Based on commit 973edd20a396e719d6847982b43cd5b0d26ae6d8 in master.

Refs #698